+++ src/components/date-picker/DatePicker.tsx
@@ -1,8 +1,8 @@
 import { useState, useMemo, useCallback } from "react"
 import { Popover, PopoverContent, PopoverTrigger } from "@/components/ui/popover"
-import { formatDate } from "@/utils/date-utils"
+import { formatDate } from "@/lib/utils/date"
 import { cn } from "@/lib/utils/utils"
-import { getDaysInMonth, getFirstDayOfMonth } from "@/utils/date-utils"
+import { getDaysInMonth, getFirstDayOfMonth } from "@/lib/utils/date"
 import { DatePickerProps } from "@/types/date-picker.types"
 import CalendarHeader from "./CalendarHeader"
 import CalendarTable from "./CalendarTable"
