+++ src/types/field-validation.ts
@@ -1,0 +1,7 @@
+export type errorType = "" | "Conflict";
+
+export interface FieldValidation {
+    error?: errorType
+    message: string
+    data?: object;
+}
